+++ examples/std/src/bin/tokio/net/send_request.rs
@@ -1,4 +1,0 @@
-// TODO: add as soon as `AsyncWebsocketClient` is implemented
-fn main() {
-    todo!()
-}