+++ src/models/transactions/escrow_finish.rs
@@ -4,8 +4,6 @@ use anyhow::Result;
 use serde::{Deserialize, Serialize};
 use serde_with::skip_serializing_none;
 
-use alloc::string::ToString;
-
 use crate::models::transactions::XRPLEscrowFinishException;
 use crate::models::{
     amount::XRPAmount,