+++ tests/common/constants.rs
@@ -1,0 +1,2 @@
+pub const ECHO_WS_SERVER: &'static str = "ws://ws.vi-server.org/mirror/";
+pub const ECHO_WSS_SERVER: &'static str = "wss://ws.vi-server.org/mirror/";